+++ cg/apps/scout/scoutapi.py
@@ -9,7 +9,7 @@ from typing import List, Optional
 import yaml
 
 from cg.apps.scout.scout_export import ScoutExportCase, Variant
-from cg.apps.scout.scout_load_config import ScoutLoadConfig
+from cg.meta.upload.scout.scout_load_config import ScoutLoadConfig
 from cg.utils.commands import Process
 
 LOG = logging.getLogger(__name__)
