+++ cg/meta/upload/scout/scout_load_config.py
